Centralise TSP tests
